refactor: extract telegram.js from main.js (BAT-197)
Move ~430 lines of Telegram Bot API code into dedicated telegram.js module:
- telegram(), telegramSendFile(), detectTelegramFileType()
- extractMedia(), downloadTelegramFile() with streaming
- cleanResponse(), toTelegramHtml(), sendMessage()
- sendTyping(), sendStatusMessage(), deleteStatusMessage(), deferStatus()
- sentMessageCache + recordSentMessage()

Depends on config.js and web.js. No logic changes.
